refactor: Move FeatureStrategies to models
